feat: swap test harness to node:test, min node 22.12.0 (#727)
BREAKING CHANGE: uses sync ESM modules, which is now possible with the release
of node 22.12.0. On top of it swaps test harness to built-in node:test.
Other major changes: support ioredis@5 - main change is lack of Bluebird promises support, which
may break plenty of code that relied on it. Updated dlock & associated modules with ioredis@5 support
Updated rdkafka to latest, which fixes a few segfaults and deadlocks on consumer close. Uses tsx to
transform typescript for tests and do proper cjs/esm interop. Jest / mocha proved to be very difficult to
implement

Short list of changes:
- removed jest
- using node:test
- removed mocha
- minimized using bluebird
- ioredis@5
- librdkaka@1.7.1
- pino@9 
- sentry@8
- minimum node @ 22.12.0

---

* feat: swap test harness to node:test, min node 22.12.0
* chore: remove jest config
* chore: remove other jest related scripts
* chore: correct types
* fix: ioredis bluebird
* fix: mocha -> node:test
* chore: update eslint plugins
* fix: migrate tests
* feat: update release-it